staging: comedi: adl_pci7x3x: fix pointer-to-int-cast warning
This driver uses the void *private variable in the comedi_subdevice
to pass the offset needed to read/write the appropriate register to
get/set the channels for the subdevice.

The adl_pci7x3x_do_insn_bits() and adl_pci7x3x_di_insn_bits() functions
were retrieving this offset by casting the s->private value as an
unsigned int. On 64-bit builds this results in a warning:

warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]

Fix these warnings by casting the void * to an unsigned long.

Reported-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Ian Abbott <abbotti@mev.co.uk>

+++ drivers/staging/comedi/drivers/adl_pci7x3x.c
@@ -121,7 +121,7 @@ static int adl_pci7x3x_do_insn_bits(struct comedi_device *dev,
 				    struct comedi_insn *insn,
 				    unsigned int *data)
 {
-	unsigned int reg = (unsigned int)s->private;
+	unsigned long reg = (unsigned long)s->private;
 	unsigned int mask = data[0];
 	unsigned int bits = data[1];
 
@@ -147,7 +147,7 @@ static int adl_pci7x3x_di_insn_bits(struct comedi_device *dev,
 				    struct comedi_insn *insn,
 				    unsigned int *data)
 {
-	unsigned int reg = (unsigned int)s->private;
+	unsigned long reg = (unsigned long)s->private;
 
 	data[1] = inl(dev->iobase + reg);
